Add PodIP ip lookup mode to Kubernetes Strategy (#82)
* Add PodIP ip lookup mode to Kubernetes Strategy

Sometimes it can be useful to get the IP of a pod directly instead of through a service. One example is when using a readiness probe that is only succesfful if the node is already connected to the cluster.

With this commit it's possible to specify a kubernetes_ip_lookup_mode of either :endpoints (default) or :pods depending on how you want to look up the ip of the pod.

Additionally this commit also adds the option of specifying the cluster name used when formatting dns records  using the mode :dns option.

* Keep current node list when kubernetes api returns error

If an error is returned from the kubernetes api the current strategy is
to return an empty list. This is unfortunate, just because the kubernetes
api has some unexpected error we shouldn't start disconnecting the
cluster.

* Add timeout to k8s api calls

Currently timeout is infinite due to default value in httpc, this is not a very useful value as it locks the update strategy in certain situations