Change Jason requirement from `~> 1.1.2` to `~> 1.1`
With the current setting, lots of people are not being able to keep all
dependencies up-to-date. Here is an example generated with `mix hex.outdated`

```
jason                         1.1.2    1.2.0   No
```

The only dependency in my application that is holding this update off is
`libcluster`. With this commit, libcluster will now allow version `1.2.0` to be
used with no need to force a dependency version via `override: true`